requirements.txt
requests
pyyaml
freesasa
biopython
pandas
numpy
tqdm
